Initial integration with Metals Scala langserver.
Of course after wiring all this up, Metals chokes and dies due to not being
able to compile the SBT/Zinc bridge internally. Presumably this is fixed in the
nightly snapshots, but we'll just wait for them to poop out an actually working
release at some point and try again. Go Team Scala.